refactor: Consolidate separate UI components in a "component" crate.
