+++ src/rust/ensogl/lib/components/src/lib.rs
@@ -17,6 +17,8 @@
 
 #![recursion_limit="512"]
 
+pub mod drop_down_menu;
+pub mod list_view;
 pub mod toggle_button;
 
 /// Commonly used types and functions.
