+++ src/rust/ensogl/lib/components/src/list_view/component.rs
@@ -1,7 +1,7 @@
 //! Select List Component
 use crate::prelude::*;
 
-use crate::entry;
+use super::entry;
 
 use enso_frp as frp;
 use ensogl_core::application;
