--- src/rust/ensogl/lib/drop-down-menu/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name    = "ensogl-gui-drop-down-menu"
-version = "0.1.0"
-authors = ["Enso Team <contact@luna-lang.org>"]
-edition = "2018"
-
-[lib]
-crate-type = ["rlib", "cdylib"]
-
-
-[dependencies]
-enso-frp             = { version = "0.1.0" , path = "../../../lib/frp"   }
-enso-logger          = { version = "0.1.1"                               }
-enso-prelude         = { version = "0.1.2"                               }
-enso-shapely         = { version = "0.1.1"                               }
-enso-types           = { version = "0.1.0" , path = "../../../lib/types" }
-ensogl-core          = { version = "0.1.0" , path = "../core"            }
-ensogl-gui-list-view = { version = "0.1.0" , path = "../list-view"       }
-ensogl-text          = { version = "0.1.0" , path = "../text"            }
-ensogl-theme         = { version = "0.1.0" , path = "../theme"           }
-
-
-[dev-dependencies]
-wasm-bindgen-test = { version = "0.3.8" }
