+++ src/rust/ide/view/graph-editor/Cargo.toml
@@ -16,8 +16,6 @@ enso-protocol             = { version = "0.1.0", path = "../../lib/enso-protocol
 enso-shapely              = { version = "0.1.1"                                              }
 ensogl                    = { version = "0.1.0", path = "../../../ensogl"                    }
 ensogl-gui-components     = { version = "0.1.0", path = "../../../ensogl/lib/components"     }
-ensogl-gui-drop-down-menu = { version = "0.1.0", path = "../../../ensogl/lib/drop-down-menu" }
-ensogl-gui-list-view      = { version = "0.1.0", path = "../../../ensogl/lib/list-view"      }
 ensogl-shape-utils        = { version = "0.1.0", path = "../../../ensogl/lib/shape-utils"    }
 ensogl-text               = { version = "0.1.0", path = "../../../ensogl/lib/text"           }
 ensogl-text-msdf-sys      = { version = "0.1.0", path = "../../../ensogl/lib/text/msdf-sys"  }
